requirements.txt
crewai>=0.102.0,<1.0.0
langchain-groq>=0.2.3,<1.0.0
yfinance>=0.2.54,<1.0.0
duckduckgo-search>=7.5.1,<8.0.0
python-dotenv>=1.0.1,<2.0.0
fastapi>=0.115.8,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
pydantic>=2.10.6,<3.0.0
